change to records for models

+++ Spendy/PriceRetrieverFactory/Interfaces/IPriceRetriever.cs
@@ -2,8 +2,8 @@
 
 public class PriceModel
 {
-    public decimal OriginalPrice { get; set; }
-    public decimal SalePrice { get; set; }
+    public double OriginalPrice { get; set; }
+    public double SalePrice { get; set; }
     public DateTime PriceDate { get; set; }
     public bool IsOnSale { get; set; }
     public bool IsAvailable { get; set; }
